nftables plugin: improve variable names about comments
There is a main distiction between comments, the one that is read from
collectd config section of this plugin, used for filtering purposes. Eg:

<Plugin nftables>
	inet filter input "in"
			  ^^^^
</Plugin>

User data comment (udata_comment) refers to the comment that is extracted from a rule
fetched using libnftnl (comments are stored inside user data of the
rule). The existance of a rule depend on the ruleset of the computer
running this.

Also, don't check for udata_comment to be NULL because free(ptr) does nothing
if ptr is NULL. nftnl_rule_get_comment will return NULL if the rule
being processed from the ruleset has no comment inside its user data.